+++ mayan/apps/smart_settings/links.py
@@ -1,3 +1,4 @@
+from django.conf import settings
 from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.navigation.classes import Link
@@ -7,6 +8,11 @@ from .icons import (
     icon_setting_edit
 )
 from .permissions import permission_settings_edit, permission_settings_view
+
+
+def condition_local_storage_enabled(context, resolved_object):
+    return not settings.COMMON_DISABLE_LOCAL_STORAGE
+
 
 link_setting_namespace_list = Link(
     icon=icon_setting_namespace_list,
@@ -25,7 +31,8 @@ link_namespace_root_list = Link(
     text=_('Namespaces'), view='settings:setting_namespace_list'
 )
 link_setting_edit = Link(
-    args='resolved_object.global_name', icon=icon_setting_edit,
+    args='resolved_object.global_name',
+    condition=condition_local_storage_enabled, icon=icon_setting_edit,
     permissions=(permission_settings_edit,), text=_('Edit'),
     view='settings:setting_edit_view'
 )